replaced ascii-encoded-string-field with cleaned-field
