+++ microcosm_flask/fields/cleaned_field.py
@@ -1,0 +1,17 @@
+"""
+CleanedField Marshmallow field
+
+An extension of the marshmallow fields.String that removes unicode control
+characters on deserialization
+"""
+
+from marshmallow import fields
+import regex
+
+
+class CleanedField(fields.String):
+    def _deserialize(self, value, *args, **kwargs):
+        if value is not None:
+            # remove control characters
+            value = regex.sub(r'\p{C}', '', value)
+        return super()._deserialize(value, *args, **kwargs)
