--- microcosm_flask/tests/fields/test_ascii_string.py
@@ -1,28 +1,0 @@
-"""
-Test language field.
-
-"""
-from hamcrest import assert_that, equal_to, is_
-from marshmallow import Schema
-
-from microcosm_flask.fields import AsciiEncodedStringField
-
-
-class AsciiStringSchema(Schema):
-    str = AsciiEncodedStringField(required=True)
-
-
-comment_text_unicode = "A nice \u0420\u043e\u0441\u0441\u0438\u044fcomment"
-comment_text_cleaned = "A nice comment"
-
-
-def test_load():
-    schema = AsciiStringSchema()
-
-    result = schema.load(dict(
-        str=comment_text_unicode,
-    ))
-    assert_that(
-        result["str"],
-        is_(equal_to(comment_text_cleaned)),
-    )
