+++ setup.py
@@ -33,6 +33,7 @@ setup(
         "python-dateutil>=2.7.3",
         "PyYAML>=3.13",
         "rfc3986>=1.2.0",
+        "regex>=2021.8.21",
     ],
     extras_require={
         "metrics": "microcosm-metrics>=2.2.0",
